Revert "[XPU][CPU] Enable the native path of DeepSeek" (#4367)
