+++ python/sglang/srt/layers/quantization/w8a8_fp8.py
@@ -14,7 +14,7 @@ from sglang.srt.layers.quantization.fp8_utils import (
     cutlass_fp8_supported,
     normalize_e4m3fn_to_e4m3fnuz,
 )
-from sglang.srt.utils import get_device_capability, is_hip
+from sglang.srt.utils import is_hip
 
 _is_hip = is_hip()
 
@@ -35,20 +35,7 @@ class W8A8Fp8Config(QuantizationConfig):
 
     @classmethod
     def get_min_capability(cls) -> int:
-        if hasattr(torch, "cuda") and torch.cuda.is_available():
-            return 89
-
-        # Vendors can update
-        return sys.maxsize
-
-    @classmethod
-    def get_availability(cls) -> bool:
-        major, minor = get_device_capability()
-        if hasattr(torch, "cuda") and torch.cuda.is_available():
-            return major * 10 + minor > 89
-
-        # Vendors can update
-        return False
+        return 89
 
     @classmethod
     def get_name(self) -> str:
