+++ python/sglang/srt/layers/quantization/w8a8_int8.py
@@ -1,4 +1,3 @@
-import sys
 from typing import Any, Callable, Dict, List, Optional
 
 import torch
@@ -19,7 +18,6 @@ from sglang.srt.layers.quantization.base_config import (
     QuantizeMethodBase,
 )
 from sglang.srt.layers.quantization.int8_kernel import per_token_quant_int8
-from sglang.srt.utils import get_device_capability
 
 
 class W8A8Int8Config(QuantizationConfig):
@@ -38,20 +36,7 @@ class W8A8Int8Config(QuantizationConfig):
 
     @classmethod
     def get_min_capability(cls) -> int:
-        if hasattr(torch, "cuda") and torch.cuda.is_available():
-            return 75
-
-        # Vendors can update
-        return sys.maxsize
-
-    @classmethod
-    def get_availability(cls) -> int:
-        major, minor = get_device_capability()
-        if hasattr(torch, "cuda") and torch.cuda.is_available():
-            return major * 10 + minor > 75
-
-        # Vendors can update
-        return False
+        return 75
 
     @classmethod
     def get_name(self) -> str:
